feat: adding flux as a cloud
This changeset adds flux as a new cloud, which largely
wraps the Kubernetes cloud class, but provides the separation
to make it clear we are deploying Flux, and to allow for other
small tweaks to the customization. This currently works to
deploy Kubernetes (via the Flux cloud) and when I uncomment
the deploy_vars "module" variable it will be able to use a
work in progress provisioner module, which is not added to
this changeset. Note that my strategy is to make as minimal
changes as possible, so I am using the same Kubernetes classes
and templates and editing only when necessary.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>