fix: use custom validation logic instead of discriminator union
The first reason is we want there be a `RequestsAggregator` base class,
so that we can use it both for sending and receiving request information.
pydantic requires a literal type field for the base class, which causes
a conflict.

The second reason is to maintain backwards compatibility. Since when using
discriminator union, we need to pass a `type` field to the Pydantic model.