detach the managed job controller from job submission
Previously, the ray driver program as well as a ray worker stayed in use for the
entire runtime of a managed job. Now, the job controller will detach from the
submitted job/ray driver and continue running in the background.

This means we have to manually manage logging as well as liveness of the
controller process. Two new directories are introduced for this purpose as well
as plumbing.