remove sky jobs launch --fast (#4467)
* remove sky jobs launch --fast

The --fast behavior is now always enabled. This was unsafe before but since
\#4289 it should be safe.

We will remove the flag before 0.8.0 so that it never touches a stable version.

sky launch still has the --fast flag. This flag is unsafe because it could cause
setup to be skipped even though it should be re-run. In the managed jobs case,
this is not an issue because we fully control the setup and know it will not
change.

* fix lint