Fetch IBM instance profiles (#5003)
Create catalog (`ibm/vms.csv`) file from the IBM instance
profiles fetched by the API.
Unfortunately, it seems that currently, it is not possible
to fetch pricing. There is the pricing API but it doesn't
work for the `kind:instance.profile`. It works for the
`kind:service` as described in the IBM documentation
(https://cloud.ibm.com/docs/account?topic=account-getting-pricing-api).

Co-authored-by: Vedran Vidović <vvidovic@croz.net>