[jobs/serve] validate controller name before updating value (#5486)
This is a follow-up to #5482. Confirmed that there is an
AssertionError with this change, before the fix in #5482. This should
make it more difficult for similar regressions to happen in the
future.

This does have the unfortunate side effect of causing an
AssertionError if you have a broken controller name and you want to
terminate it, because you will hit the AssertionError during sky down.
I think this is fine. If people hit that error, which should be very
rare, we can ask them to downgrade, down the cluster, and then upgrade
again.