fix service account user_id length
This was intended to be 16 random chars, but due to how
secrets.token_hex works, it was actually 32, for a total user_id
length of 35 chars (too long!).

This was causing SSH to give the error `unix_listener: path
"/tmp/skypilot_ssh_blahblahblah" too long for Unix domain socket`.
Since we construct the ControlPath including the user_id, the length
was contributing to the problem.

Confirmed that we don't see the unix_listener error after shortening
the SA user_id.