[test] fix test_skyserve_dynamic_ondemand_fallback failure (#6545)
Fixes #6539.

Before, `--infra` flag would not override all of cloud/region/zone.
After #6392, it will override all three, even if only one is
specified. (So `--infra gcp` means "GCP, any region", even if there is
a region specified in the YAML.)

Since the region info is specified in the YAML, the CLI `--infra gcp`
flag is not needed. So we just remove it.