[test] make test_log_collection more robust
The existing test will not just grep the actual job output, but also
some setup log lines. In some case, those log lines can contain all
the digits 1-9 as well as "10", which would cause the test to
erroneously pass. This occured in #7135 with
test_log_collection_on_gcp, because this log line was present:
```
2025-09-17 21:53:46,296\tINFO worker.py:1719 -- Connected to Ray cluster. View the dashboard at \u001b[1m\u001b[32m127.0.0.1:8266 \u001b[39m\u001b[22m
```
Which matches all numbers from 1 to 10.