[jobs] catch some errors from get_job_status
It's possible that backend.get_job_status throws a TypeError or
ValueError in some rare cases. This just indicates that the cluster
was cleaned up in the middle of the status check, probably by the
skypilot-status-refresh-daemon. This is fine, and we should just treat
the cluster is down and recover the managed jobs.

We should gracefully handle these errors within
backend.get_job_status, but this will make sure that managed jobs
aren't broken in the mean time.