[jobs] limit the max number of jobs that can run/launch (#7488)
* [jobs] limit the max number of jobs that can run/launch

Previously, after #7051, we allowed the number of jobs
launching/running to scale purely based on the controller resources.

After this change, we set maximum values for the number that can run
and launch, since there are some bottlenecks (e.g. DB) that will not
necessarily scale with the instance resources.

* address comments