[deps] make sure server deps are always installed
Before this change, `pip install skypilot[aws]` could not correctly
run a local API server in some cases, because the server dependencies
were not installed.

We should have a way to install just the skypilot client without any
of the server deps (#5500), but we don't have that at the moment.