[debug] add details when managed jobs state transition fails (#7594)
* [debug] add details when managed jobs state transition fails

Before, we would just get a cryptic "0 rows updated", but it wasn't
clear why - usually because the job is in the wrong state, but what
state?

Note that this approach isn't 100% race-proof as the status could
change between the first and second query, but it seems like a big
improvement anyway.

* add unit test

* fix lint issues

* refactor