[jobs] improved logging (#7595)
* [jobs] improved logging

Remove the logic for creating a custom logger object for each job
launched from the jobs controller process. Instead, rely on the
hijacked contextual sys.stdout/stderr implementations that are used by
the API server for logging in async contexts.

Also, add some additional logs to make it easier to see which
controller process is related to which job, and some other
improvements.

* fix log flushing

* fix context overrides

* use the context manager

* fix test

* make the test fast

* lint